refactor: 重构项目结构
更新 .gitignore，排除编译产物 main.js 与 styles.css，避免将
构建输出直接提交到仓库，改为通过发行版发布以保持代码库纯净。

在 esbuild 配置中增加对 CSS 输出文件重命名的插件：
在构建结束时将生成的 main.css 重命名为 styles.css，确保构建产
物命名符合项目预期并与 .gitignore 中的忽略规则一致；同时引入
fs 与 path 模块以实现文件操作。

移除项目中的 styles.css 源文件（内容已由构建流程生成或不再需
要），减少源代码冗余并依赖构建产物进行样式生成与发布。
