+++ .gitignore
@@ -8,6 +8,10 @@
 # npm
 node_modules
 
+# Don't include the compiled main.js file in the repo.
+# They should be uploaded to GitHub releases instead.
+main.js
+styles.css
 # Exclude sourcemaps
 *.map
 
@@ -17,7 +21,5 @@ data.json
 # Exclude macOS Finder (System Explorer) View States
 .DS_Store
 
-main.js
-
-# other
+# Environment variables file
 .env
