+++ main.ts
@@ -1,1 +1,2 @@
-export {default} from './src/ExportImagePlugin';
+export { default } from './src/ExportImagePlugin';
+import './styles/style.css';
