+++ src/components/file/ModalContent.css
@@ -1,0 +1,5 @@
+.export-image-preview-left {
+    .setting-item {
+        padding: 10px 0;
+    }
+}
