+++ src/types/settings.ts
@@ -1,4 +1,4 @@
-import { isCreatable } from './imageFormatTester';
+import { isCreatable } from '../imageFormatTester';
 
 // eslint-disable-next-line @typescript-eslint/naming-convention
 export const DEFAULT_SETTINGS: ISettings = {
